reactor: Split into `reactor` with the core update and warn checking code, and `reactor_thread` with all the threading stuff.
